implemented searching and sorting
redux for the sorting?... could dispatch an action to set the strategy in the store
sorting would then happen in the store logic and not here
would feel better to me right now and I will probably implement it like that soon

Especially since the "display worst score" feature will be coming soon. Would be nice just to dispatch an event and having the right leaderboard entrys fall out of the selector without much logic in the component.